exampleofgoofbackend/POVBackend-prod/streaming-env/requirements.txt
aioboto3==13.1.1
aiobotocore==2.13.1
aiofiles==24.1.0
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
b2sdk==2.4.1
boto3==1.34.131
botocore
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
jmespath==1.0.1
logfury==1.0.1
multidict==6.0.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
typing_extensions==4.12.2
urllib3==1.26.19
watchdog==4.0.1
wrapt==1.16.0
yarl==1.9.4
redis==5.0.7
Flask==3.0.3
psutil==5.8.0